Responded to review, ready to merge
